+++ index.html
@@ -35,7 +35,7 @@
     <link rel="stylesheet" href="./scss/main.scss" />
     <!-- animation    -->
 
-    <script src="./js/animation.js" type="module"></script>
+    <script src="./js/index.js" type="module"></script>
 
     <!-- some extensions -->
   </head>
