--- js/test.js
@@ -1,3 +1,0 @@
-console.log(new Date("2022-07-19T19:14").getTime() / 1000);
-console.log(new Date("2022-07-19").getTime() / 1000);
-console.log(Date.now() / 1000);
